Bug fixes and remove useless file.

+++ Sources/ARNETWORK_IOBuffer.c
@@ -177,7 +177,7 @@ void ARNETWORK_IOBuffer_Delete(ARNETWORK_IOBuffer_t **IOBufferPtrAddr)
     }
 }
 
-eARNETWORK_ERROR ARNETWORK_IOBuffer_AckReceived(ARNETWORK_IOBuffer_t *IOBufferPtr, int seqNumber)
+eARNETWORK_ERROR ARNETWORK_IOBuffer_AckReceived(ARNETWORK_IOBuffer_t *IOBufferPtr, uint8_t seqNumber)
 {
     /** -- Receive an acknowledgement to a IOBuffer -- */
 
@@ -185,7 +185,7 @@ eARNETWORK_ERROR ARNETWORK_IOBuffer_AckReceived(ARNETWORK_IOBuffer_t *IOBufferPt
     eARNETWORK_ERROR error = ARNETWORK_OK;
 
     /** delete the data if the sequence number received is same as the sequence number expected */
-    if(IOBufferPtr->isWaitAck && IOBufferPtr->seqWaitAck == seqNumber)
+    if((uint8_t)IOBufferPtr->isWaitAck && IOBufferPtr->seqWaitAck == seqNumber)
     {
         IOBufferPtr->isWaitAck = 0;
         error = ARNETWORK_IOBuffer_PopDataWithCallBack(IOBufferPtr, ARNETWORK_MANAGER_CALLBACK_STATUS_ACK_RECEIVED);
